Endpoints

Modify and examine defined open311 endpoints. o311_endpoints() retrieves a list of endpoints including additional information. o311_add_endpoint adds to this list to define a new endpoint that can be used for queries. o311_reset_endpoints restores the initial state of the endpoints list.

Usage

o311_add_endpoint(
  name,
  root,
  jurisdiction = NULL,
  key = FALSE,
  pagination = FALSE,
  limit = NULL,
  json = TRUE
)

o311_reset_endpoints()

o311_endpoints()

Arguments

name

[character]

Name of an endpoint / city. This name can be arbitrary and only serves for identification in [o311_api].

root

[character]

Base URL of the endpoint for sending production-grade requests. The root URL commonly points to "georeport/v2/".

jurisdiction

[character]

Unique identifier of the jurisdiction. The jurisdiction is typically defined as the domain of the respective city website. It is optional as most endpoints only serve one jurisdiction.

key

[logical]

Is an API key mandatory?

pagination

[logical]

Are requests responses paginated?

limit

[integer]

If paginated, how many requests does one page contain?

json

[logical]

Are JSON responses supported? If FALSE, defaults to "XML" responses. "XML" responses are harder to simplify.

Value

For o311_endpoints, a dataframe containing all relevant information on an endpoint. For o311_add_endpoint, the new endpoint, invisibly. o311_reset_endpoints returns NULL invisibly.

Details

o311_endpoints() returns a static list defined in the package installation directory. This list contains a limited number of endpoints that were proven to work at the time of package development. It does not include newer/smaller/less known endpoints or test APIs. These can be manually added using o311_add_endpoint.

Note

This function uses R_user_dir to persistently store custom endpoints data between sessions. To clean up, run o311_reset_endpoints() which deletes the package-specific user directory and defaults back to system.file("endpoints.json", package = "open311").

See also

Examples

# read default endpoints
o311_endpoints()
#> # A tibble: 18 × 8
#>    name                    root  docs  jurisdiction key   pagination limit json 
#>    <chr>                   <chr> <chr> <chr>        <lgl> <lgl>      <int> <lgl>
#>  1 Annaberg-Buchholz, Deu… http… http… annberg-buc… FALSE TRUE          50 TRUE 
#>  2 Bloomington, IN         http… NA    bloomington… FALSE TRUE        1000 TRUE 
#>  3 Bonn, Deutschland       http… http… bonn.de      FALSE TRUE         100 TRUE 
#>  4 Boston, MA              http… http… boston.gov   FALSE TRUE          50 TRUE 
#>  5 Brookline, MA           http… http… brooklinema… FALSE TRUE          50 TRUE 
#>  6 Chicaco, IL             http… http… cityofchica… FALSE TRUE          50 TRUE 
#>  7 Köln / Cologne, Deutsc… http… http… stadt-koeln… FALSE TRUE          50 TRUE 
#>  8 Ottawa, ON              http… NA    ottawa.ca    FALSE TRUE        1000 TRUE 
#>  9 Rostock, Deutschland    http… http… rostock.de   FALSE FALSE         NA TRUE 
#> 10 San Francisco, CA       http… http… sfgov.org    FALSE TRUE          50 TRUE 
#> 11 Siegburg, Deutschland   http… http… siegburg.de  FALSE TRUE          50 TRUE 
#> 12 Turku, FI               http… NA    turku.fi     FALSE FALSE         NA TRUE 
#> 13 Zürich, CH              http… http… stadt-zueri… FALSE TRUE         100 TRUE 
#> 14 Philadelphia, PA        http… NA    philadelphi… FALSE TRUE          NA TRUE 
#> 15 Gilbert, AZ             http… http… gilbertaz.g… FALSE TRUE          20 TRUE 
#> 16 Schwerin, DEU           http… http… schwerin.de  FALSE FALSE         NA TRUE 
#> 17 FixMyStreet             http… http… fixmystreet  FALSE TRUE        1000 TRUE 
#> 18 München, DEU            http… http… muenchen.de  FALSE TRUE         100 TRUE 

# add a new endpoint
o311_add_endpoint(name = "test", root = "test.org/georeport/v2")

# read new endpoints
o311_endpoints()
#> # A tibble: 19 × 8
#>    name                    root  docs  jurisdiction key   pagination limit json 
#>    <chr>                   <chr> <chr> <chr>        <lgl> <lgl>      <int> <lgl>
#>  1 Annaberg-Buchholz, Deu… http… http… annberg-buc… FALSE TRUE          50 TRUE 
#>  2 Bloomington, IN         http… NA    bloomington… FALSE TRUE        1000 TRUE 
#>  3 Bonn, Deutschland       http… http… bonn.de      FALSE TRUE         100 TRUE 
#>  4 Boston, MA              http… http… boston.gov   FALSE TRUE          50 TRUE 
#>  5 Brookline, MA           http… http… brooklinema… FALSE TRUE          50 TRUE 
#>  6 Chicaco, IL             http… http… cityofchica… FALSE TRUE          50 TRUE 
#>  7 Köln / Cologne, Deutsc… http… http… stadt-koeln… FALSE TRUE          50 TRUE 
#>  8 Ottawa, ON              http… NA    ottawa.ca    FALSE TRUE        1000 TRUE 
#>  9 Rostock, Deutschland    http… http… rostock.de   FALSE FALSE         NA TRUE 
#> 10 San Francisco, CA       http… http… sfgov.org    FALSE TRUE          50 TRUE 
#> 11 Siegburg, Deutschland   http… http… siegburg.de  FALSE TRUE          50 TRUE 
#> 12 Turku, FI               http… NA    turku.fi     FALSE FALSE         NA TRUE 
#> 13 Zürich, CH              http… http… stadt-zueri… FALSE TRUE         100 TRUE 
#> 14 Philadelphia, PA        http… NA    philadelphi… FALSE TRUE          NA TRUE 
#> 15 Gilbert, AZ             http… http… gilbertaz.g… FALSE TRUE          20 TRUE 
#> 16 Schwerin, DEU           http… http… schwerin.de  FALSE FALSE         NA TRUE 
#> 17 FixMyStreet             http… http… fixmystreet  FALSE TRUE        1000 TRUE 
#> 18 München, DEU            http… http… muenchen.de  FALSE TRUE         100 TRUE 
#> 19 test                    test… NA    NA           FALSE FALSE         NA TRUE